ensure workspace folder is the working directory
I had the 'mainBeanFile' setting set to my main beancount file, 'main.bean', but the other files were being checked as though the setting was blank. It turns out that the existsSync call was returning false because the current working directory was "/". This change ensures that the current working directory is set to the first folder in the workspace, so the filename reference in 'mainBeanFile' is relative. This could have been because I'm running on a Mac (MacOS 10.13.4)? Anyway, it works for me now.